+++ utils/ai/context/dictionary.ts
@@ -6,25 +6,7 @@ import {
   type Channel,
   type GuildBasedChannel,
 } from "discord.js";
-
-export interface SerializedAttachment {
-  id: string;
-  name: string;
-  url: string;
-  mime_type: string | null;
-  size: number;
-  width: number | null;
-  height: number | null;
-}
-
-export interface ReferencedMessage {
-  id: string;
-  author_id: string;
-  content: string;
-  timestamp: string;
-  referenced_message_id: string | null;
-  attachments?: SerializedAttachment[];
-}
+import type { ReferencedMessage } from "./types";
 
 export class ContextDictionary {
   private users = new Map<string, User | GuildMember>();
